requirements/api.txt
Flask==2.2.*
Flask-Cors==3.0.10
neo4j==4.4.1
newspaper3k==0.2.*
requests==2.27.*
spacy==3.3.0
numpy==1.21
lark==1.1.2
gunicorn==20.1.0
dill==0.3.5.1
importlib-metadata==4.13.0
celery[redis]==5.2.7
networkx==2.6.*
backoff==2.2.*
elasticsearch==8.5.1